implementasi axios dan blade template

+++ app/Http/Controllers/API/BookController.php
@@ -3,37 +3,132 @@
 namespace App\Http\Controllers\API;
 
 use App\Http\Controllers\Controller;
-use Illuminate\Support\Facades\Validator;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use Symfony\Component\HttpKernel\Exception\HttpException;
 use App\Models\Book;
 use OpenApi\Annotations as OA;
+use Validator;
 
 /**
- * Class BookController.
- *
- * @author Josevan <josevan422024005@ukrida.ac.id>
+ * Class Controller,
+ * 
+ * @author Josevan <josevan422024005@civitas.ukrida.ac.id>
  */
-
 class BookController extends Controller
 {
-    /**
+    /** 
      * @OA\Get(
      *     path="/api/book",
      *     tags={"book"},
-     *     summary="Display a listing of items",
+     *     summary="Display a listing of the items",
      *     operationId="index",
      *     @OA\Response(
      *         response=200,
-     *         description="Successful",
+     *         description="successful",
      *         @OA\JsonContent()
-     *     )
+     *     ),
+     *  @OA\Parameter(
+     *      name="_page",
+     *      in="query",
+     *      description="current page",
+     *      required=true,
+     *      @OA\Schemas(
+     *          type="integer",
+     *          format="int64",
+     *          example=1
+     *      )
+     *  ),
+     *  @OA\Parameter(
+     *      name="_limit",
+     *      in="query",
+     *      description="max item in a page",
+     *      required=true,
+     *      @OA\Schemas(
+     *          type="integer",
+     *          format="int64",
+     *          example=10
+     *      )
+     *  ),
+     *  @OA\Parameter(
+     *      name="_search",
+     *      in="query",
+     *      description="word to search",
+     *      required=false,
+     *      @OA\Schemas(
+     *          type="string"
+     *      )
+     *  ),
+     *  @OA\Parameter(
+     *      name="_publisher",
+     *      in="query",
+     *      description="search by publisher like name",
+     *      required=false,
+     *      @OA\Schemas(
+     *          type="string"
+     *      )
+     *  ),
+     *  @OA\Parameter(
+     *      name="_sort_by",
+     *      in="query",
+     *      description="word to search",
+     *      required=false,
+     *      @OA\Schemas(
+     *          type="integer",
+     *          example="latest"
+     *      )
+     *  ),
      * )
      */
-    public function index()
+    public function index(Request $request)
     {
-        $book = Book::all();
-        return response()->json(array('message' => 'Data retrieved successfully', 'data' => $book), 200);
+        try {
+            $data['filter'] = $request->all();
+            $page = $data['filter']['_page'] = (@$data['filter']['_page'] ? intval($data['filter']['_page']) : 1);
+            $limit = $data['filter']['_limit'] = (@$data['filter']['_limit'] ? intval($data['filter']['_limit']) : 1000);
+            $offset = ($page ? ($page - 1) * $limit : 0);
+            $data['products'] = Book::whereRaw('1 = 1');
+
+            if ($request->get('_search')) {
+                $data['products'] = $data['products']->whereRaw('(LOWER(title) LIKE "%' . strtolower($request->get('_search')) . '%")');
+            }
+            if ($request->get('_type')) {
+                $data['products'] = $data['products']->whereRaw('LOWER(type) = "' . strtolower($request->get('_type')) . '"');
+            }
+            if ($request->get('_sort_by')) {
+                switch ($request->get('_sort_by')) {
+                    default:
+                    case 'latest_publication':
+                        $data['products'] = $data['products']->orderBy('publication_year', 'DESC');
+                        break;
+                    case 'latest_added':
+                        $data['products'] = $data['products']->orderBy('created_at', 'DESC');
+                        break;
+                    case 'name_asc':
+                        $data['products'] = $data['products']->orderBy('title', 'ASC');
+                        break;
+                    case 'name_desc':
+                        $data['products'] = $data['products']->orderBy('title', 'DESC');
+                        break;
+                    case 'price_asc':
+                        $data['products'] = $data['products']->orderBy('price', 'ASC');
+                        break;
+                    case 'price_desc':
+                        $data['products'] = $data['products']->orderBy('price', 'DESC');
+                        break;
+                }
+            }
+            $data['products_count_total'] = $data['products']->count();
+            $data['products'] = ($limit == 0 && $offset == 0) ? $data['products'] : $data['products']->limit($limit)->offset($offset);
+            // $data['products_raw_sql']       = $data['products']->toSql();
+            $data['products'] = $data['products']->get();
+            $data['products_count_start'] = ($data['products_count_total'] == 0 ? 0 : (($page - 1) * $limit) + 1);
+            $data['products_count_end'] = ($data['products_count_total'] == 0 ? 0 : (($page - 1) * $limit) + sizeof($data['products']));
+            return response()->json($data, 200);
+
+        } catch (\Exception $exception) {
+            throw new HttpException(400, "Invalid data : {$exception->getMessage()}");
+        }
     }
 
     /**
@@ -57,13 +152,13 @@ class BookController extends Controller
      *         description="Request body description",
      *         @OA\JsonContent(
      *             ref="#/components/schemas/Book",
-     *             example={"title": "Eating Clean", "author": "Inge Tumiwa-Bachrens",
-     *                 "publisher": "Kawan Pustaka", "publication_year": "2016",
-     *                 "cover": "https://images-na.ssl-images-amazon.com/images/S/compressed.photo.goodreads.com/books/1482170555l/3351107.jpg",
-     *                 "description": "Menjadi sehat adalah impian semua orang. Makanan yang selama ini kita pikir sehat ternyata belum tentu ‘sehat’ bagi tubuh kita.",
-     *                 "price": 85000}
-     *         )
-     *     )
+     *             example={"title": "Eating Clean", "author": "Inge Tumiwa-Bachrens", "publisher": "Kawan Pustaka", "publication_year": "2016",
+     *                      "cover": "https://images-na.ssl-images-amazon.com/images/S/compressed.photo.goodreads.com/books/1482170055i/33511107.jpg",
+     *                      "description": "Menjadi sehat adalah impian semua orang. Makanan yang selama ini kita pikir sehat ternyata belum tentu 'sehat' bagi tubuh kita.",
+     *                      "price": 85000}
+     *         ),
+     *     ),
+     *      security={{"passport_token_ready":{}, "passport":{}}}
      * )
      */
     public function store(Request $request)
@@ -73,18 +168,18 @@ class BookController extends Controller
                 'title' => 'required|unique:books',
                 'author' => 'required|max:100',
             ]);
-
             if ($validator->fails()) {
                 throw new HttpException(400, $validator->messages()->first());
             }
-
             $book = new Book;
             $book->fill($request->all())->save();
-            return response()->json(array('message' => 'Saved successfully', 'data' => $book), 200);
+            return $book;
+
         } catch (\Exception $exception) {
-            throw new HttpException(400, "Invalid data - {$exception->getMessage()}");
+            throw new HttpException(400, "Invalid Data : {$exception->getMessage()}");
         }
     }
+
     /**
      * @OA\Get(
      *     path="/api/book/{id}",
@@ -115,14 +210,18 @@ class BookController extends Controller
      *             type="integer",
      *             format="int64"
      *         )
-     *     )
+     *     ),
      * )
      */
     public function show($id)
     {
-        $book = Book::findOrFail($id);
-        return response()->json(array('message' => 'Data detail retrieved successfully', 'data' => $book), 200);
+        $book = Book::find($id);
+        if (!$book) {
+            throw new HttpException(404, 'Item not found');
+        }
+        return $book;
     }
+
     /**
      * @OA\Put(
      *     path="/api/book/{id}",
@@ -159,33 +258,38 @@ class BookController extends Controller
      *         description="Request body description",
      *         @OA\JsonContent(
      *             ref="#/components/schemas/Book",
-     *             example={"title": "Eating Clean", "author": "Inge Tumiwa-Bachrens",
-     *                 "publisher": "Kawan Pustaka", "publication_year": "2016",
-     *                 "cover": "https://images-na.ssl-images-amazon.com/images/S/compressed.photo.goodreads.com/books/1482170555l/3351107.jpg",
-     *                 "description": "Menjadi sehat adalah impian semua orang. Makanan yang selama ini kita pikir sehat ternyata belum tentu ‘sehat’ bagi tubuh kita.",
-     *                 "price": 85000}
-     *         )
-     *     )
+     *             example={"title": "Eating Clean", "author": "Inge Tumiwa-Bachrens", "publisher": "Kawan Pustaka", "publication_year": "2016",
+     *                      "cover": "https://images-na.ssl-images-amazon.com/images/S/compressed.photo.goodreads.com/books/1482170055i/33511107.jpg",
+     *                      "description": "Menjadi sehat adalah impian semua orang. Makanan yang selama ini kita pikir sehat ternyata belum tentu 'sehat' bagi tubuh kita.",
+     *                      "price": 85000}
+     *         ),
+     *     ),
+     *      security={{"passport_token_ready":{}, "passport":{}}}
      * )
      */
     public function update(Request $request, $id)
     {
-        if (!$id) {
-            throw new HttpException(400, "Invalid id");
-        }
-
         $book = Book::find($id);
         if (!$book) {
             throw new HttpException(404, 'Item not found');
         }
 
         try {
+            $validator = Validator::make($request->all(), [
+                'title' => 'required|unique:books',
+                'author' => 'required|max:100',
+            ]);
+            if ($validator->fails()) {
+                throw new HttpException(400, $validator->messages()->first());
+            }
             $book->fill($request->all())->save();
-            return response()->json(array('message' => 'Updated successfully', 'data' => $book), 200);
+            return response()->json(array('message' => 'Updated successfully'), 200);
+
         } catch (\Exception $exception) {
-            throw new HttpException(400, "Invalid data - {$exception->getMessage()}");
+            throw new HttpException(400, "Invalid Data : {$exception->getMessage()}");
         }
     }
+
     /**
      * @OA\Delete(
      *     path="/api/book/{id}",
@@ -203,7 +307,7 @@ class BookController extends Controller
      *         @OA\JsonContent()
      *     ),
      *     @OA\Response(
-     *         response=204,
+     *         response=200,
      *         description="Successful",
      *         @OA\JsonContent()
      *     ),
@@ -216,15 +320,23 @@ class BookController extends Controller
      *             type="integer",
      *             format="int64"
      *         )
-     *     )
+     *     ),
+     *      security={{"passport_token_ready":{}, "passport":{}}}
      * )
      */
     public function destroy($id)
     {
-        $book = Book::findOrFail($id);
-        $book->delete();
+        $book = Book::find($id);
+        if (!$book) {
+            throw new HttpException(404, 'Item not found');
+        }
 
-        return response()->json(array('message' => 'Deleted successfully', 'data' => $book), 204);
+        try {
+            $book->delete();
+            return response()->json(array('message' => 'Deleted successfully'), 200);
+
+        } catch (\Exception $exception) {
+            throw new HttpException(400, "Invalid data : {$exception->getMessage()}");
+        }
     }
-
 }
